Fix failing blind transfer test after PJSIP 2.4 upgrade.
A change in PJSIP 2.4 resulted in the blind transfer test taking
longer than 30 seconds to complete. On my box, the test on PJSIP 2.3
tood 22 seconds, and on PJSIP 2.4 took 38 seconds.

The reason for this was due to some changes to transaction timers in
PJSIP. In our test scenario, alice would send an initial INVITE with
branch "foo". Asterisk would answer the INVITE, and alice would ACK the
answer. In PJSIP 2.3, the transaction was immediately destroyed when the
ACK was received. However, in 2.4, there is a delay between when the ACK
is received and when the transaction is terminated.

Alice re-used branch "foo" in subsequent requests. With PJSIP 2.3, this
was not an issue since the previous transaction had been destroyed
already. With 2.4, however, this incoming request would match the
previous transaction, and the request would be ignored. After the
original transaction timed out and was destroyed, then the next
retransmission of the request from alice would finally be responded to.

The solution for this is to vary the branch parameter that we send in
requests from alice in this scenario. This is done by adding a numerical
suffix to the end of the [last_Via:] to ensure that the branch will be
different from previous transactions.

ASTERISK-25147 #close
Created by Mark Michelson

Change-Id: Ifdf798cb7abd49f764d533b9af96aa813f954391