channels/pjsip/subscriptions/rls - Improve logging
The primary purpose of this is to improve logging in the RLS tests so that
the exact phases of the test and expectations being evaluated are known
at any given time when reviewing logs. In addition, some improvements have
been made to the integrity checks so that failures are reported back to the
rls_test module.

ASTERISK-25144
Reported by: Jonathan Rose

Change-Id: Iea3f5bb8a424d7ff813a85d6c8224a14ebc3e0dd