pjsip: Replace usage of localhost with 127.0.0.1
The upgrade to pjproject 2.4 on the Bamboo agents caused these to
start failing as no valid resolver is available. Since these tests
aren't really testing the resolver they don't need to be using
localhost.

A separate task for creation of a fallback system level resolver
has been created for fixing the resolver side of things.

Change-Id: I0c24ab8bec33823bfec4f5f0e567dc920605542f