pjsip: Add resolver test for NAPTR record order.
This test ensures that the NAPTR record with the lowest order value is
used when multiple records with different order values exist. If the
record with the lowest order value is used, the TCP transport will be
used. Otherwise the UDP transport will be used. A SIPp instance only
listens on TCP and therefore the test will only pass if the call is sent
over TCP.

ASTERISK-25012 #close
Reported by: Matt Jordan

Change-Id: I751fdcb9120fc7af2c44d041a96e4aa550e979cc