Fix test: callee initiated attended transfer w/Replaces, direct media, hold
The test has been bouncing due to an additional occurrence of an event.
This should fix the test from bouncing by allowing more than one
occurrence of the event while only performing the hold on the first
occurrence.

This also clarifies the docstring regarding the 'trigger-on-count'
option for the event action module.

Change-Id: I44d232dc281ac4b4b136f8dba0f578dc5de591d8