Add NAPTR test for preference.
This test is very similar to the NAPTR order test. The difference is
that the NAPTR records in this test have the same order but different
preference values. We ensure that the record with the lowest preference
value is chosen.

A SIPp scenario listens on a TCP transport. If the NAPTR record with the
lower preference value is used, then the resolution result should end up
sending the INVITE to this SIPp scenario. If the other NAPTR record is
chosen, then Asterisk will end up erroneously sending on a UDP
transport, resulting in the SIPp scenario timing out.

ASTERISK-25013 #close
Reported by: Matt Jordan

Change-Id: Iabed2613bc934becb3c8e6381e4af1634aa553be