Add NAPTR preference failover test.
This test has two NAPTR records. The one with the lower preference value
points to a TCP SRV record that does not exist. The one with the higher
preference value points to a UDP SRV record that will resolve properly.

A SIPp scenario is running on a UDP port and receives the call since
Asterisk fails over to UDP after the TCP SRV lookup fails.

ASTERISK-25015 #close
Reported by Matt Jordan

Change-Id: I000cd05d489c98c6b49ad9690d40d0db697b52da