Add test for SIP BYE with Also header
This test launches ten simultaneous SIPp scenarios. Each scenario places
a call to an extension that calls a local channel. The local channel
then calls the Echo() application.

The SIPp scenarios send a BYE with Also header in order to transfer the
;1 side of the local channel to the Echo() application as well.

This test exists both to ensure that a SIP BYE with Also header
functions as expected as well as to ensure that an observed deadlock
that had previously occurred does not happen any longer.

ASTERISK-25139

Change-Id: I687d28a725980fb6ccbc9b8a361d2603baee522c