testsuite: Address RLS test failures
pyxb was reliant upon core xml functionality that was overridden on
our build agents by PyXML. Since the version of Twisted that the
build agents use relies upon PyXML, we can't simply discard that, so
we are unable to use pyxb on the build agents for the time being.
This patch replaces pyxb with generated code that uses lxml to parse
the XML instead.

ASTERISK-25144 #close
Reported by: Jonathan Rose

Change-Id: I17583e5da602853e2e1814df8c712a59c4bbb288