cdr/cdr_dial_subroutines: Add a check for a parallel dial in a subroutine
This patch updates the cdr_dial_subroutines test to also cover a
parallel dial initiated from a subroutine. From a Macro, we now perform
a parallel dial of two Local channels. The first Local channel proceeds
as it previously did, with a one second Wait to help ensure ordering of
the CDR records as they are dispatched. The second Local channel hits a
Busy and hangs up. This covers the bug exposed noted in ASTERISK-24443.

ASTERISK-24443

Change-Id: Ib918fa0d0906f895dfaa4deff64535482bbe183f