pjsip: Add blonde transfer tests and connected line update checks.
Both the caller and callee initiated nominal local attended transfer tests
have been updated to include checking for connected line updates on
completion of the attended transfer. If they do not occur or the values do not
match who the other side should be the test will now fail.

Blonde attended transfer tests have also been added. A blonde transfer is
when an attended transfer is completed before the new called party has
actually answered. These tests mirror that of the nominal case and ensure
that channels are bridged as expected and that connected line updates
occur.

ASTERISK-25157

Change-Id: I5aa4aee05ae4bbd3e6fc110e608459b4c102b725