Stasis attended transfers: No StasisEnd event for transfer target channel
This required a change in Asterisk since the transfer target was not being
hung up, thus no event. However, these tests were altered to now not expect
a test failure. This patch also fixes a few typos/copy paste errors found
in the test that were keeping them from passing. As well, in a couple of
tests the local attended transfer channels needed to be hung up. In the
'non_stasis_app_to_stasis_bridge' the hangup of 'bob' was moved to be done
once the local channel entered the bridge. This made the test a little
more deterministic.

With the changes it was verified that the tests pass and that Asterisk gets
shutdown gracefully (i.e. is not killed).

ASTERISK-24782 #close
Reported by: John Bigelow
ASTERISK-25167
Reported by: Corey Farrell

Change-Id: Ife584d91394ff69a3662004471421761789c4901