Add test for rewriting contact on request with Record-Routes.
This test has a SIPp scenario call Asterisk. The SIPp scenario
represents a proxy that is between Asterisk and the actual endpoint
that placed the call towards Asterisk. As a result, the INVITE that
Asterisk receives has Record-Route headers in it. We ensure that,
despite the fact that the rewrite_contact option is enabled for the
endpoint on which the INVITE is received, that the contact is not
actually rewritten since the first item in the route set is the
same address from which the request is received. We test this by
inspecting the BYE request that Asterisk sends to SIPp to end
the call.

Change-Id: Ia6ca09a65d59e80f0a7f3e85eed946bd4564222c