Stasis & Attended transfers: Bouncing test due to off nominal path
Sometimes it is possible for the transferer to send a 'bye' to the transferee
immediately after the refer has been received. If this happens the transferee
gets hung up and the transfer does not go through (this is a valid situation).
Modified a couple of tests so that if the expected transfer events are not
received the test still passes.

ASTERISK-24782
Reported by: John Bigelow

Change-Id: I9aee48b2dee47ad422caebdd4b861d25d5a0f75b