tests/ARI: test for getting all module information
Test that returns a list of all modules that are to be
loaded based on the modules.conf configuration. The test
checks the response body of the modules, so that it matches
that of the expected response body specified within the yaml
configuration. It also checks to make sure that a 200 OK was
received back as a response from the request.

* Receives list of all loaded modules

ASTERISK-25173
Reported By: Matt Jordan

Change-Id: Ic16508cf03548a55ffa0339b94d432c7d8f660c1