Stasis & Attended transfers: Bouncing test due to off nominal path
Initially it was thought that the reason several of the stasis/attended
transfer tests were bouncing/failing was due to the endpoint hanging up
too quickly and stopping the transfer process midstream. However, that
was not the case. It was actually due to a race condition between the
initiating transfer thread and the bridge channel impart thread. Since
that is the case this tests can now expect all events to be received.

ASTERISK-24782
Reported by: John Bigelow

Change-Id: I9aee48b2dee47ad422caebdd4b861d25d5a0f75b