lib/python/asterisk/version: Correct erroneous error message
The a Git version string is parsed, the handling for the revision
portion incorrectly set 'self.handled' to be True, where the actual
variable being checked is 'handled'. This caused an error message to
erroneously be raised on a version string parsed from an Asterisk Git
checkout. This patch corrects that, and also adds a basicLogger for the
unit tests so that they print out the ERROR message as well.

Change-Id: I3e9bbefe81bf0edc24f0b427a3830bbaec6fbc16