Test/ARI: Test for getting a single module
Test that returns a single requested module that is to be
autoloaded based on the modules.conf configuration. The test
checks the response body of the module, so that it matches
that of the expected response body specified within the yaml
configuration. It also checks to make sure that a '200 OK' was
received back as a response from the request.

* Receives a single loaded module

ASTERISK-25173

Change-Id: I3dd06fccddac0086879360f9569172662a897af1