Convert Lua RFC 2833 DTMF test to Python.
This is a straightforward conversion to use the SIPp test case and the
AMI event module in order to convert the Lua test to Python. The biggest
changes were that the SIPp scenario had to have paths for PCAP files
altered in order for them to be opened properly.

ASTERISK-25225
Reported by Matt Jordan

Change-Id: I2f00155be9810bbeffb69820b5b055c736305bb4