tests/channels/pjsip: Update tests expected results for 'media_encryption'
The 'media_encryption' setting does not understand the default value of
'none', even though that it is its default value. This can cause
problems if a configuration actually sets the value to 'none'. When this
was fixed in Asterisk, the various tests that check default values of
things weren't updated. This patch updates those tests so that they pass
again.

Change-Id: I40a0a31229b77edff6e68345a9943c92900b3428