rest_api/asterisk/config/dynamic/create: Make test tolerant of interference
The res_pjsip_config_wizard module uses an in-memory sorcery wizard to
provide caching. If this module is loaded, it will cause the object
creation test to fail, as an object that should *not* be created will be
created due to the presence of the in-memory cache. This patch tells
this test to noload the res_pjsip_config_wizard module.

Change-Id: I20d9bec751426dccb31e4ff592dc820c2a93f593