pjsip: Handle initial NOTIFY in MWI aggregate test.
The PJSIP MWI aggregate test currently assumes that all NOTIFY
messages each pjsua receives will follow the expected results.
While this was true in the past this is no longer true. This is
due to the act of registration causing an initial NOTIFY to be
sent. What makes this problem worse is that the callback is
registered after the SIP registration is done, causing a window
where the initial NOTIFY may or may not be received by the test.

This change makes it so that if a NOTIFY is received that does
not match the expected result and the expected result is the first
one the NOTIFY is ignored. Since the test requires that all results
be received and that they come in order this works fine. If
unexpected NOTIFY messages continue to be received while expecting
the first result the test will eventually time out and fail.

ASTERISK-25278 #close

Change-Id: If5382e91575a771324a1462489c416f0e4e3d8ee