Testsuite/ARI: Added a pluggable module for log validation.
This pluggable module is used to verify a log files existence. It
will look inside the current test's log directory and check to see
if the given file is present or not. Another option, pass-if-present,
can be set to True or False based on what we expect the test to do.

* Added log validation pluggable module
* Log files can be verified in the testuite

ASTERISK-25252

Change-Id: I3bf8e63071e579552838e814aeda0ba5d6bb6180