SIP: Rewrite tcpauthtimeout test in python
What was formerly a single Lua test has been split into two python
tests, though the functionality of the tests is identical.

timeout_should_happen: In this test, arbitrary data is written to the
SIP TCP socket in Asterisk, and we ensure that the connection is dropped
after the configured time in sip.conf

timeout_should_not_happen: This is a simple invocation of a SIPp UAC
scenario. The call lasts longer than the configured auth timeout, but
since the call is successfully established, the timeout should be
irrelevant. The test passes if the SIPp scenario passes.

ASTERISK-25225
Reported by Matt Jordan

Change-Id: I4413f70bfb06110e8f1ba61041e0769214f2a539