bridge_wait: Check for RINGING_INBAND test event.
A recent change to Asterisk 13 and master made local channels incapable
of changing from the "UP" state to the "RINGING" state. The
bridge_wait_e_options test expected the local channel that entered the
BridgeWait() application to change state to "RINGING" as a way of
ensuring that the e(r) option worked as expected. However, since that
state change is now invalid, the event that is checked needs to be
altered, too. This changeset has the test expect the RINGING_INBAND test
event instead.

ASTERISK-25292
Reported by Kevin Harwell

Change-Id: If33efe8161b139e8c0c79690bb6eb6cf189cca9b