dialog_info_xml: Fix SIPp scenario variable names and regex.
The SIPp scenario contains multiple regular expressions and assigns the same
variable name for each. This causes the regex checks following the first check
to always be successful even if they do not match. This changes the variable
names so they are all unique.

Fixing the variable names resulted in two expressions failing which appear to
be issues with the regex and not Asterisk. This additionally fixes the regex.

A minor change to the test-config.yaml was made to add TEST_FRAMEWORK as a
dependency for the test since it expects a TestEvent.

Change-Id: I9a380bc36e243fc0a3dcc1dfc1db999e7308099a