SIP: Rewrite tcpauthlimit test in Python
What was formerly a Lua test has been ported to Python such that the
test can now be executed from the Jenkins build agents. Functionally,
the test is identical to its predecessor.

The test is composed of two types of scenarios: SIP Client and SIPp.

The SIP Client scenarios attempt to create n+1 TCP socket connections to
Asterisk, where n is the value of the tcpauthlimit property in sip.conf.
If the tcpauthlimit property is honored, the (n+1)th socket connection
will fail.

The SIPp scenarios attempt to create n*2 SIPp processes, where n is the
value of the tcpauthlimit property in sip.conf. Each SIPp scenario is
configured to connect to the same Asterisk host. If the tcpauthlimit
property is honored, only n of these scenarios will pass, while the
remaining n will fail.

ASTERISK-25225
Reported by Matt Jordan

Change-Id: Ica28ba0ca7ae92b3546da4cd23458f289c111d36