pjsip: Handle initial NOTIFY in MWI unsolicited mailbox count changes test.
The PJSIP MWI unsolicited mailbox count changes test currently
assumes that all NOTIFY messages pjsua receives will follow
the expected results. While this was true in the past this is
no longer true. This is due to the act of registration causing
an initial NOTIFY to be sent. What makes this problem worse is
that the callback is registered after the SIP registration is
done, causing a window where the initial NOTIFY may or may not
be received by the test.

This change makes it so that if a NOTIFY is received that does
not match the expected result and the expected result is the first
one the NOTIFY is ignored. Since the test requires that all results
be received and that they come in order this works fine. If
unexpected NOTIFY messages continue to be received while expecting
the first result the test will eventually time out and fail.

ASTERISK-25298 #close

Change-Id: Ie8d5a46c350497fc77de5eff34c4bef7c00bbdd3