tests/channels/SIP/sip_channel_params: Convert to Python test
This change converts the tests/channels/SIP/sip_channel_params test
to be Python based instead of Lua based. As part of this the way
the test fundamentally works has been changed to be consistent with
other tests and to also reduce the test execution time. Instead of
relying on AMI to query for values this is done as part of the
dialplan execution. Once all values have been checked the test ends
immediately.

Change-Id: I15ce51ddb9bc72ecab37771414b5ab81c7d44c4d