bridge: Add test for confirming hangup after feature timeout
This test creates a chain of Local channels and sets up a
scenario where DTMF is sent that causes a feature timeout to
occur. When the DTMF is received at the far end the channel is
hung up. All channels are confirmed to have been hung up when
this happens.

ASTERISK-25341 #close

Change-Id: Icfa456ea9de7b452c57fa91919ca03c56120d523