lib/python/asterisk: Add a pluggable module that checks for hangups via ARI
Some ARI tests may be slightly non-deterministic as to when they end. In those
cases, it is often nice to watch channels and - if all channels have hung up -
end the test. This patch adds such a pluggable module, which mimics a similar
pluggable module that does the same using AMI.

Note that since there is no ordering between AMI and ARI events, we need to
have one for ARI for those tests that are explicitly looking to validate ARI.

Change-Id: I7ca5e5f8d5fb50503bac78d2a0fd47a981f1d142