tests/channels/pjsip/dialplan_functions: Add support for testing PJSIP Call-ID
Under ASTERISK-25352, the SIP Call-ID of the INVITE request that created the
PJSIP channel can now be extracted using the CHANNEL function. This patch
updates the test to look for the presence of a Call-ID using that function.
Given that the Call-ID is randomly generated, we simply test that we can
get any string back when looking for that item.

ASTERISK-25352

Change-Id: If8ed77cb16db11dc31880080e1c0c1d0c35bbd35