pjsip rls large notify test - fails to send notify due
This test was not sending the notify because of an unsupported transport error.
What's happening was there was no transport associated with the contact, so it
would default to UDP and since the test was configured for TCP and not UDP it
failed to find a transport. Adding a transport parameter to the contact header
ensures the use of the correct transport.

Also, even though this test failed to send the notify it was still passing
because the test simply timed out with no failure condition. To fix this a
failure condition was also added that checks to make sure a particular event
is received, and if it isn't then it will fail.

ASTERISK-25363

Change-Id: I1a61a7efd72bfe8681b684510433b57df7b9c5a8