lib/python/asterisk: Use old style inheritance call
In 98b77f234b03a8644fd71aec685a2a9db78fbab3, the parent object initializer
for the AriClientProtocol class was invoked for Autobahn 0.10.5-2+. However,
prior to that version (or at least on some older versions), the parent classes
were 'old-style' classes that did not inherit from Object. As a result, using
the new style invocation mechanism of 'super' causes an exception to be thrown
with older versions of Autobahn.

This patch uses the direct invocation method, thereby (hopefully) supporting
both new and old versions of Autobahn.

Change-Id: Ic3bde314d8f6634e2bf58e46b39f963c82bac50f