tests/channels/pjsip/dialplan_functions: Update tests for dynamic contacts
Recently, a change went into the PJSIP stack in Asterisk that modified storage
of dynamic contacts in PJSIP to use a MD5 hash. This was done so that
relatively arbitrarily long contact URIs could be stored and retrieved. This
patch updates two tests that used the old formatting for dynamic contacts,
pjsip_aor and pjsip_contact, to expect the new MD5 hash.

Change-Id: Ib62fe519136e552020542d34cd1440e915ceaf49