Testsuite: tests/channels/SIP/tcpauthlimit/tcp_client_scenario is Failing
While this does not actually correct the issue with the intermittent
failures, this patch modifies a log message such that it matches the
formatting of the other log messages (to help with debugging) and
also changes the use of LineReceiver.transport.loseConnection to
LineReceiver.transport.abortConnection. Reviewing the test logs
revealed that not all of the transports were being destroyed after a
test case executed. After reading the Twisted documentation closer,
it appears that LineReceiver.transport.loseConnection will have no
effect under certain circumnstances. So, in situations where you do
not need to take advantage of the safeguards provided by:
LineReceiver.transport.loseConnection, the method to use is:
LineReceiver.transport.abortConnection.

http://twistedmatrix.com/documents/current/core/howto/servers.html#
loseconnection-and-abortconnection

ASTERISK-25432
Reported By: Ashley Sanders

Change-Id: I8f6d448fb15f43245644c1c9c416dffe47944a15