Attended transfer tests timing out, but still passing
The nominal caller and callee local attended transfer tests were timing out due
the referer waiting on a BYE from Asterisk when it should have actually been
sending one instead. Changed the scenarios to send a BYE.

Also fixed a problem in the CoordinatedScenario class where a deferred callback
was not being called with the right amount of parameters.

Change-Id: I2bd79cf9af911fc1b8fb4f454a4dd06cac55a2bf