lib/python/asterisk/ari: Only attempt to init the parent of AriClientProtocol
Newer versions of Autobahn require the derived classes of
WebSocketClientProtocol to initialize the base classes. However, older versions
of Autobahn use old style Python classes without an initializer. As a result,
the only way to handle this situation is to attempt to initialize the parent
class(es), and catch the error. Since initializing an old style class with a
call to super will always raise a TypeError, we opt for that approach here.

Change-Id: I334d7c8dacb9ab5b3d7f3b918b6e43476d9c69c1