chan_sip and res_pjsip transports trying to bind to same port
There still exists some tests (for instance, callee_local_blonde) that attempt
to load both chan_sip and res_pjsip modules at the same time and bind to their
transports to the same port.

The current conflicts file was "no loading" chan_sip for tests with a specified
dependency on res_pjsip. However, some tests opt to specify chan_pjsip as the
dependency instead. This patch makes it so if either or are listed as a
dependency then chan_sip won't load.

Change-Id: Ib4500a161e34f56d835975d90d82e3bd67a1a61f