tests/hep/pjsip-ipv6: Be tolerant to different IPv6 nomenclatures
An IPv6 address may include the brackets '['/']' or it may omit them.
It is possible for Asterisk to change (at a minimum) the connection address
in the SDP to use one nomenclature over another depending on whether or not
the address has been cached recently. As a result, the regex handling the
200 OK would sporadically fail.

This patch makes the checking of the IPv6 addresses more tolerant to both
nomenclatures. It does this on values that *may* be produced by Asterisk in
packets received from Asterisk by the SIPp scenario. This should make the test
more reliable in the long run.

Change-Id: I747d45fc6537f9d4f92c62da2c6a44c758a05e9d