reinvite_early: Verify operation when get an unexpected reINVITE.
A SIPp scenario places a call into Asterisk.  Once the call has been
answered, the caller sends a reINVITE before sending the ACK to the
initial INVITE.  Asterisk should respond to the reINVITE with a 491
because it has not seen the ACK completing the initial INVITE transaction.
Once the reINVITE is completed the SIPp scenario completes the original
INVITE and hangs up the call.

The reason this test exists is because Asterisk would crash on the early
reINVITE because the reINVITE was not identified and was treated as an
initial INVITE.  The call wound up with two PBX threads running on the
same channel.  This test ensures that the code changes have fixed the
issue as expected.

ASTERISK-25404
Reported by Chet Stevens

Change-Id: Id6fcdb655228b9bc89764cff00fdb921e6571152