Confbridge test failures due to scenarios sometimes ending too early
A few confbridge tests would sporadically fail because the scenario(s) would
end before all events had been processed. This happened because an "end"
scenario action was tied to a specific event, so once that event was received
it triggered the scenario to end. The test would end, but sometimes there
were available events still left to process.

This patch makes it so the test scenario will not end until all expected events
have been received, or a timeout occurs.

Change-Id: I3e067f7cd80568c9b507ca850771040d5fe2f593