Testsuite: batched rls subscription failure
While this does not actually correct the issue with the test failure (which
incidentally turns out is not localized to this test), this patch modifies the
the test driver such that the test output produces log messages that help to
elimanate it as the suspect in the test failure. Also, this patch includes
fixes for pylint errors (e.g. missing docstrings).

This is part one of an n-patch series of refactorings to help determine the
root cause of the test failure and correct pylint issues.

ASTERISK-25430
Reported By: Ashley Sanders

Change-Id: I15c005ec6a43919e5fe533c8f2dc8fb3d5c41e64