Testsuite: batched rls subscription failure
While this does not actually correct the issue with the test failure (which
incidentally turns out is not localized to this test), this patch modifies the
the test harness by renaming the test to a consistent name with respect to
its module. The issue was in that there were two classes in two modules, one
was the harness, the other was the validator, but the names were subverted and
it was getting difficult to keep them straight when debugging.
Also, this patch includes missing or more thorough docstrings such that the
intent of the class and its methods is more obvious.

This is part two (2) of an n-patch series of refactorings to help determine the
root cause of the test failure and correct pylint issues.

ASTERISK-25430
Reported By: Ashley Sanders

Change-Id: Ie3af00440aa011786afd3241bae27d39c3518079