Testsuite: batched rls subscription failure
While this does not actually correct the issue with the test failure (which
incidentally turns out is not localized to this test), this patch modifies the
the test config structure such that the full_state and resources collections
are now represented by a single collection, packets. Having them disconnected
in the configuration was obscuring the intent.

This is part three (3) of an n-patch series of refactorings to help determine
the root cause of the test failure and correct pylint issues.

ASTERISK-25430
Reported By: Ashley Sanders

Change-Id: I3bc60588b1430d152de608a0c2a733a843411666