asterisk/sipp: Add option to SIPpScenarioSequence to control stopping reactor
By default, SIPpScenarioSequence will stop the TestCase object reactor when the
final SIPp scenario in the sequence has finished. In some cases, however,
particularly when the SIPp scenarios are being used to drive a more complex set
of tests, this isn't desireable. Instead, the test is stopped based on more
complex criteria outside of the SIPp scenario's knowledge.

This patch adds a new option to the SIPpScenarioSequence constructor that
informs it to stop the reactor when all scenarios have executed. By default,
this option is set to True for backwards compatibility reasons.

Change-Id: I4811f25ec1a9c77bbd122de43e4e228d7f5a9c2a