Testsuite: Allow CDR test to specify CDR directory
On one of our internal test builds, cdr_csv is unavailable but we
still want to perform a CDR record check using the same parameters as
what we would expect from cdr_csv. Fortunately, cdr_custom can be used
to write a CDR log identical to what cdr_csv writes. However, since the
test class for CDR matching was written with cdr_csv in mind, we can't
tell it to use the cdr-custom folder. This patch will allow the test
yaml that defines the CDR match to specify which folder the CDR file
resides in.

Change-Id: I8148499a447a27129ca6620effcfc3bf7b40a0d4