AMI: Support a reconnecting AMIFactory if it exists
A PR has been up against starpy that would make the AMIFactory reconnect
on a loss of connection. This is useful if we ever want to test restarts of
Asterisk. However, for this to occur, the consumers of the AMIFactory must be
able to be aware that a reconnect occurred so that they can drop their
reference to the AMIProtocol and obtain the new one.

This patch makes it so that ami/test_case will attempt to use the reconnecting
version of the factory by providing an on_reconnect handler when creating the
AMIFactory. If it does not exist or throws an exception, they fall back to the
older version of AMIFactory.

Change-Id: Ifc9becf8ae263dc594f3b3600f6ea09948dd0749