tests/remote-test: Add a sample remote test
This patch adds a test that only executes if the Test Suite is in 'remote'
mode. It will attempt to connect to a remote Asterisk instance, and spawn a
Local channel at s@default. If the channel is answered, the test will pass.

This test is provided merely as an example of how to write a remote test.

Change-Id: I857c5f20d4c260be73697062506f63cb3533c3a8