runtests: Add some additional attributes to the JUnit XML report
This patch adds a few new attributes/nodes to the JUnit XML report. This
includes:
(1) Skipped tests, including the reason why they were skipped. This is useful
    for knowing why a test wasn't run on a build agent.
(2) The time the test run was started.

Additionally, the 'dry-run' attribute has been removed, as it is illegal within
a JUnit compliant schema. Instead. if the dry-run option is selected, all tests
are counted as being skipped.

Change-Id: Id95d7eb379c35ce8be563a28c5828fd76c0c1387