runtests: Add the ability to run the test suite in a loop
This patch adds a new option to runtests.py, '-n'. This will instruct the
test suite to run its set of tests 'n' number of times. Each result from
an iteration of the test suite is stored as a new <testsuite> node in the
resulting JUnit XML report. If a negative number is provided for the '-n'
option, the Test Suite will run continuously until told to stop.

In order to tell the Test Suite to stop, signal handlers have been added for
the USR1 and USR2 signals.
 - If the USR1 signal is sent to the python process, the Test Suite will stop
   executing any further tests after the current test case completes. Any
   non-executed tests are marked as skipped.
 - If the USR2 signal is sent to the python process, the current executing
   test case is marked as abandoned and failed. Any subsequent tests in the
   current running set of tests are marked as skipped.

This prevents us from having to use a keyboard interrupt to stop the Test
Suite, losing all results.

Change-Id: I6c65250a45e48dc36b7c4657d4cfbe63efeb1a6d