Testsuite: batched rls subscription failure
While this does not actually correct the issue with the test failure (which
incidentally turns out is not localized to this test), this patch modifies the
the test validator by separating the concerns for each validation method,
improving the logging, and fixing pylint issues.

This is part four (4) of an n-patch series of refactorings to help determine the
root cause of the test failure and correct pylint issues.

ASTERISK-25430
Reported By: Ashley Sanders

Change-Id: Icb8af439f5c3db3efb0a93d6402ab4dcacdf3731