Adds option to send crash reports via email
--email-on-crash enables email on crash behavior. When this
behavior is enabled, the testsuite will attempt to email crash
reports (currently just the path to the test and a backtrace).
Configuration for sending the email (smtp server, who to send as,
who to send to, etc) is handled via crash-mail-config.yaml which
goes in the base folder for the testsuite when in use. There is
a sample configuration for it stored in the sample-yaml folder
(crash-mail-config.yaml.sample)

Change-Id: I17def41f15294a48754269d9c42aa649389840ad