Testsuite: Fix sporadic failures in attended transfer tests.
These tests would sometimes fail due to the SIPp scenarios receiving an
unexpected BYE while expecting a COLP update. The problem was that the
channel for charlie & alice (SIPp #1) would be hung up to quickly after
the transfer. This would cause a BYE to be sent to the alice channel
before the expected COLP update would occur.

This fixes the problem by triggering the hang up upon the COLP updates
occurring rather than the channel hang ups as a result of the transfer.

This does not fix the failure originally reported as it could not be
reproduced.

ASTERISK-25536 #close
Reported By: Kevin Harwell

Change-Id: I15413539945424cf1ff604f2b0cfa7ede94acaf8