testsuite: Allow log files to be disabled by tests.
The testsuite currently always creates a full log file with
debug and higher log output as well as a messages log file with
info and higher. For tests that generate a lot of output this
can overwhelm python itself. This change allows tests to disable
either of these log files if they want.

The con of doing so is that you have to rely on the Asterisk log
files to understand what is going on instead.

ASTERISK-25562 #close

Change-Id: I55924b1f68373204617f26dd2ac8c1cef5ae4468