channels/pjsip/qualify/qualify_timeout: Increase sipp wait time.
Some agents can be quite slow at times causing things to not
exactly execute when they are expected to. In the case of the
qualify_timeout test there was little time left between marking a
contact as unreachable and receiving the 200 OK. This change
increases the time to give Asterisk more time to mark it as
unreachable.

Change-Id: I7dd0591453ebce18514894bf46ed23285afc6705